Move RAWR-related classes and functions into rawr.py.
